adjusting image conversion

+++ pdfmaker_preprocessing.rb
@@ -149,6 +149,7 @@ images = Dir.entries("#{Bkmkr::Paths.project_tmp_dir_img}").select {|f| !File.di
 image_count = images.count
 corrupt = []
 processed = []
+skipped = []
 
 if image_count > 0
   FileUtils.cp Dir["#{Bkmkr::Paths.project_tmp_dir_img}/*"].select {|f| test ?f, f}, pdftmp_dir
@@ -157,30 +158,36 @@ if image_count > 0
     puts i
     pdfimage = File.join(pdftmp_dir, "#{i}")
     imgfilename = i.split(".").shift
+    imgformat = i.split(".").pop
     jpgimage = File.join(pdftmp_dir, "#{imgfilename}.jpg")
-    if i.include?("fullpage")
-      #convert command for ImageMagick should work the same on any platform
-      `convert "#{pdfimage}" -colorspace gray "#{jpgimage}"`
-      unless pdfimage.include?("jpg")
-        Mcmlln::Tools.deleteFile(pdfimage)
-      end
-      filecontents = filecontents.gsub(/#{pdfimage}/,jpgimage)
-      processed << pdfimage
-    elsif i.include?("_FC") or i.include?(".txt") or i.include?(".css") or i.include?(".js")
+    if i.include?("_FC") or i.include?(".txt") or i.include?(".css") or i.include?(".js")
       FileUtils.rm("#{pdfimage}")
-    else
-      myres = `identify -format "%y" "#{pdfimage}"`
-      if myres.nil? or myres.empty? or !myres
-        corrupt << pdfimage
+    end
+    unless imgformat == "jpg" or imgformat == "jpeg" or imgformat == "png" or imgformat == "pdf" or imgformat == "ai"
+      if i.include?("fullpage")
+        #convert command for ImageMagick should work the same on any platform
+        `convert "#{pdfimage}" -colorspace gray "#{jpgimage}"`
+        unless pdfimage.include?("jpg")
+          Mcmlln::Tools.deleteFile(pdfimage)
+        end
+        filecontents = filecontents.gsub(/#{pdfimage}/,jpgimage)
+        processed << pdfimage
       else
-        resize = calcImgSizes(myres, pdfimage, maxheight, maxwidth, grid)
-        `convert "#{pdfimage}" -density #{myres} #{resize}-quality 100 -colorspace gray "#{jpgimage}"`
+        myres = `identify -format "%y" "#{pdfimage}"`
+        if myres.nil? or myres.empty? or !myres
+          corrupt << pdfimage
+        else
+          resize = calcImgSizes(myres, pdfimage, maxheight, maxwidth, grid)
+          `convert "#{pdfimage}" -density #{myres} #{resize}-quality 100 -colorspace gray "#{jpgimage}"`
+        end
+        unless pdfimage.include?("jpg")
+          Mcmlln::Tools.deleteFile(pdfimage)
+        end
+        filecontents = filecontents.gsub(/#{pdfimage}/,jpgimage)
+        processed << pdfimage
       end
-      unless pdfimage.include?("jpg")
-        Mcmlln::Tools.deleteFile(pdfimage)
-      end
-      filecontents = filecontents.gsub(/#{pdfimage}/,jpgimage)
-      processed << pdfimage
+    else
+      skipped << pdfimage
     end
   end
 end
@@ -234,6 +241,10 @@ File.open(Bkmkr::Paths.log_file, 'a+') do |f|
       f.puts "IMAGE PROCESSING ERRORS:"
       f.puts "The following images encountered processing errors and may be corrupt:"
       f.puts corrupt
+  end
+  if skipped.any?
+    f.puts "Skipped processing the following images:"
+    f.puts skipped
   end
   f.puts "Uploaded the following images to the ftp:"
   f.puts ftpstatus
